07-项目实战/智能金融投研助手-RAG/requirements.txt
# 向量模型（embedding + reranker）
sentence-transformers>=3.0.0

# 检索与索引
rank-bm25>=0.2.2
faiss-cpu>=1.8.0

# 中文分词
jieba>=0.42.1

# PDF 解析（可选，当前使用 sample_blocks.json）
# pymupdf>=1.24.0
# pdfplumber>=0.11.0

# RAG 编排（可选，Agentic RAG 项目使用）
# langgraph>=0.2.0

# 评估（可选）
# ragas>=0.2.0
